+++ examples/foxes/compare_wakes/run.py
@@ -76,7 +76,7 @@ if __name__ == "__main__":
         "-w",
         "--wakes",
         help="The wake models",
-        default=["Bastankhah_linear_k004", "PorteAgel_linear_k004"],
+        default=["Bastankhah2014_linear_k004", "Bastankhah2016_linear_k004"],
         nargs="+",
     )
     parser.add_argument(