+++ examples/foxes/sequential/run.py
@@ -32,7 +32,7 @@ if __name__ == "__main__":
         "-w",
         "--wakes",
         help="The wake models",
-        default=["Bastankhah_linear_lim_k004"],
+        default=["Bastankhah2014_linear_lim_k004"],
         nargs="+",
     )
     parser.add_argument(